android build: Add `androidx.swiperefreshlayout` dependency.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@99ef72cad. This must happen
at or before the main upgrade commit because the API is apparently
needed in the new version. (It can happen before because it should
be harmless to declare a dependency even if we don't use it.)

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2