flipper (android): Prepare to enable Flipper (1/5).
Done as part of the RN v0.61 -> v0.62 upgrade.

In this commit:
- Part of the RN v0.60 -> v0.61 [sic] changes to the template
  app [1], corresponding to facebook/react-native@3a66fc7dc
- A partial reversion of that, in the RN v0.61 -> v0.62 changes to
  the template app [2], corresponding to
  facebook/react-native@bb272bacc.

It looks like this was prematurely released in RN v0.61 instead of
RN v0.62 [3], the latter being where most sources claim Flipper
support was added; a list of such sources is on CZO [4].

In that partial reversion, the initialize-Flipper function's only
call site is removed.

We'll re-enable Flipper after the main upgrade commit, not before,
just to be safe [5]. That will correspond to
facebook/react-native@05f5cb534.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.60.6&to=0.61.5
[2] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2
[3] https://github.com/zulip/zulip-mobile/issues/3782#issuecomment-642248809
[4] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Flipper/near/900499
[5] https://github.com/zulip/zulip-mobile/pull/4244#discussion_r482536188