compose [nfc]: Cut now-resolved comment.
The issue this describes was filed as zulip/zulip#16176 and has now
been fixed on the server, in zulip/zulip@6ee9042d4.  This code
doesn't do anything about the issue, because fortunately it doesn't
need to -- the symptom is harmless (indeed the webapp had the same
behavior, and that went unnoticed for some time.)  So we can skip
having a comment bring it to the reader's attention.

The "This is meant to align" comment applies to this whole stretch
of logic, so we can delete the blank line to group it all together.