styles/tabs types: Use actual `ViewStyle` type.
Modern versions of react-navigation have `style` typed better than
`{ ... }`, so we might as well too.

It's odd that we have to add some type-casts; the values at those
properties that may be clobbered by `...style` are subtypes of the
corresponding properties on `ViewStyle`. Anyway, we'll be able to
remove them following the RN v0.61 -> v0.62 upgrade, when we get
Flow v0.113.

This also lets us remove a flowlint suppression; not sure why.