flow: Stop pointing to old location of a RN-internal file.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to the two deletions in the Flow config in
facebook/react-native@e54ecf907.

We pointed to the file's new location in a recent commit before the
main upgrade commit; now, we finish following
facebook/react-native@e54ecf907 by removing the references to the
old, now-incorrect location of the file.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2