accountsReducer tests: Clarify expectations and data flow in REALM_ADD.
Each of these started by constructing a piece of the expected result
state, and then went on to construct the input.  That works but gets
confusing to read, as the logical flow of information zigzags up and
down the test code.  Instead, describe first the input and then the
expected result.

Meanwhile in the "move to front of list" test, as the name suggests,
what we're looking for is for the matching account information to
move to the front; so say that directly instead of constructing a
fresh example account.  There are nuances when the zulipVersion or
zulipFeatureLevel needs an update, but we cover those in their own
test cases just below, and the fresh `eg.makeAccount` call didn't
really add any clarity about them.