accountsReducer tests: Simplify on REALM_INIT by writing data in line.
No need to give names to `newState`, `expectedState`, etc., when
they're each used just once and their roles can be communicated
directly by where they appear in the `expect` expression.

Also no need to check `expect(newState).not.toBe(prevState)`;
we're already checking it's equal to `expectedState`, and the
latter is indeed not equal to `prevState`.