navSelectors: Get data from `NavigationService`, not from Redux.
We haven't been using any of these functions in react-redux
`connect` calls; if we had, we might have been relying on them for
the important task of triggering updates (re-`render`s) in some
components, when the nav state changes.

We've recently set up the `NavigationService`, which should provide
the nav state just as readily as Redux could.

Using React Navigation's types, instead of our own, reveals an
unchecked assumption we've been making, at the single call site of
`getChatScreenParams`. The value at `.narrow` in the returned
object, if that value is not undefined, won't obviously be a
`Narrow`. So, add a FlowFixMe there.