messagesActions: Don't let `doNarrow` cancel an expected navigation (2/2).
In the status quo, in all cases where `doNarrow` cancels a
navigation, it's a bug -- the user's attempt to navigate is being
silently thwarted, which causes confusion.

So, remove the existing check for whether the Redux store is
hydrated.

As Greg points out [1], this removal is fine because the act of
navigating (to a screen that will crash if the store is not
hydrated) happens in a thunk action (`doNarrow`) that will be held
up in the queue managed by the middleware from `redux-action-buffer`
until rehydration is complete.

We're confident that a thunk action will be buffered like this
because the thunk middleware comes after the `redux-action-buffer`
middleware in `listMiddleware`; see 42adfb348 for more detail on
that design.

The actual plain-object navigation action (created by `StackActions`
in navActions.js), if it were dispatched directly, may not be
buffered [2], because the `react-navigation-redux-helpers`
middleware comes before the `redux-action-buffer` middleware. But
we're safe as long as the code that dispatches it is inside
`doNarrow` or another thunk action creator.

[1] https://github.com/zulip/zulip-mobile/pull/4273#discussion_r497981247

[2] This didn't stand up to some quick testing, but some caution
    wouldn't be harmful. In any case, when we eliminate the
    `react-navigation-redux-helpers` middleware and separate the
    navigation-action dispatch from Redux entirely,
    `redux-action-buffer` will have no control over when the
    navigation happens.