i18n: Cut a string we don't actually use, "Narrow to conversation".
The last place we said "Narrow to conversation" in the UI was
removed in 2fb9d2ee9, in 2018.  Should have removed it here at the
same time, but apparently we forgot.

Noticed this while looking to see if we use the term "narrow"
anywhere in the UI.  In general we probably don't want to do so; its
meaning isn't particularly transparent, and we don't systematically
use it across the UI in order to teach it to users.  The one place
I'm aware that we use it in the Zulip UI in general is in the help
modal in the webapp for its keyboard shortcuts, to explain what the
shortcuts S and Shift+S and their friends do.