ios: Update launch screen.
Use a "Storyboard" instead of an "XIB", as has apparently [1] been
required since 2020-04-30. A change like this is part of the RN
v0.62 -> v0.63 changes to the template app [2], corresponding to
facebook/react-native@33b3a1a14.

Since it's simple to do (especially after adding the "Brand" color
set in a recent commit), make it be a solid brand-color screen
instead of solid white. It should match the color of our
`LoadingScreen`, which is often the first thing that displays when
you open the app.

The changes in this commit coincide with the iOS-specific
instructions from `expo-launch-screen` for how to set up a launch
screen [3], with some exceptions:

- We don't actually go on to set up `expo-launch-screen`; we
  currently don't intend to [4].

- We don't do anything marked "optional".

- We don't add an image to the launch screen; we can do that later
  if we want to.

- We set the `Storyboard ID` to `LaunchScreenViewController`, not
  `SplashScreenViewController`. Apple regularly calls the thing a
  "launch screen", not a "splash screen".

[1] https://developer.apple.com/news/?id=03042020b
[2] https://react-native-community.github.io/upgrade-helper/?from=0.62.2&to=0.63.3
[3] https://github.com/expo/expo/tree/master/packages/expo-splash-screen#manual-configuration
[4] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/React.20Navigation.20v5/near/1042794
[5] https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/launch-screen/