reducer tests [nfc]: Say `prevState` instead of `initialState`.
We've got a mix of `prevState` and `initialState` for what we call
the state before we've run a test action through the reducer. It
would be good to be consistent.

Between them, `prevState` is probably better for accuracy; the
variable is not used to store the state at the *very beginning*, but
the state just previous to something interesting happening (i.e.,
just before the reducer processes an action). And the reducer files
already use `initialState` for something that *does* mean the state
at the very beginning, e.g. (in usersReducer),

```
const initialState: UsersState = NULL_ARRAY;
```