notif: Normalize set of users to represent a group PM thread.
This fixes #4293 from the other direction as we did in the previous
commit.  As a result we avoid the at-least-latent bug where we can
end up with duplicate narrows in our Redux state; and we also avoid
any unknown other bug where, like with #4293, we're relying on the
set of users having been filtered this way.