outbox: Replace narrowContains with something more solid.
For conversation narrows -- individual topics, and PM conversations
-- this relied completely on our various representations of the
narrow completely agreeing on the order of operators and operands.
For PM conversations it also relied on the set of users identified
(including or excluding self) and the order they're listed in.

This causes several bugs, and exacerbates possible others:

 * We have a number of different ways those users can be ordered in a
   group PM conversation, and each of those means a bug here.  One of
   those is #4298.

 * Until quite recently, we also had longstanding bugs where the set
   of users would vary, and we may well still have more.  This meant
   an additional symptom for each of those bugs.

 * Because we rely on comparing two `JSON.stringify` results for
   equality, we're implicitly assuming that objects we want to
   consider equal always have their properties enumerate in the same
   order.  In particular, we regularly say `operator` first and then
   `operand`; if there's some way we can also end up with `operand`
   first and then `operator`, then this code wouldn't see the
   resulting narrow as equal to narrows it should be equal to.  That
   makes this at least a latent bug, as there's nothing else in the
   code expressing this assumption (even in comments) and it would
   have been easy to break.

This code also had a dumb bug where if the haystack was a whole-stream
narrow, we'd blindly assume the needle was a stream or topic narrow.
If you had an outbox message addressed to a PM conversation, and then
managed to visit a stream whose name looked like your correspondent's
email address (or their addresses, comma-separated), then you'd see
the outbox PM in that stream narrow.

This commit fixes all those issues by replacing the use of this
function with `isMessageInNarrow`, which avoids all the above
assumptions.

More broadly: the question we're actually asking at this function's
one caller is whether a specific *outbox message* belongs in a given
narrow.  It's sort of an accident that the inclusion relation on
*narrows* even mostly corresponds to that question.  Better to use
a function that's explicitly about our actual question.

Fixes: #4298