outbox: Stop duplicating self as PM recipient; sort all recipients.
This fixes a bug affecting sending self-PMs.  The bug's current
symptom is that when you hit send, the outbox message shows in the
message list... with its own sender header with name and avatar, even
when it's right after another message (which necessarily has the same
sender, namely yourself.)  If the send happens quickly, this shows up
mostly as a flicker glitch, where the other messages move briefly up
and then back down, because the message is taller with the sender.

The root issue is that when constructing the display_recipient of an
outbox PM, we were indiscriminately appending the self user after
taking the list of users given in the Narrow value.  Presumably this
was motivated by the fact that display_recipient values for PMs list
all participants including self, while for most PM narrows we leave
out the self user... but for self-PMs, we don't.  The result is that
for self-PMs we were producing a malformed display_recipient,
consisting of the self listed twice.

The observed symptom then comes because in `renderMessages`, we see
the outbox message has a different recipient from the previous one,
and treat it as if we expected to show a recipient bar for it.  (We
don't actually show a recipient bar, but that causes us to show the
sender header.)

Fix the issue by only adding the self user if it's not already in the
list.  While we're here, also sort the recipients in a canonical way,
by ID.

Fixes: #4301