narrow [nfc]: Add keyFromNarrow to encapsulate JSON.stringify.
We'd like to replace the Narrow type in most of our code with a new,
better-structured one.  In general this kind of refactoring gets a lot
of help from the type-checker: you change the type of a value in one
place, and Flow points out all the other places that consume or
provide that value and need to be updated to match.

But one thing it won't catch is where we pass a Narrow to
JSON.stringify, as we do all the time.  That's because the latter will
take any value at all as its argument (the type in the signature is
`mixed`) -- so passing a value of some new type is still well-typed,
but it breaks the code's implicit assumption that the value had the
structure of Narrow.

We can fix this by encapsulating this usage of JSON.stringify inside a
function that specifically expects a Narrow.  Here's that function.
In the next commit, we'll switch over all our uses of JSON.stringify
that should use it.