msglist: Don't assume people with the same name are the same.
Discovered this bug by reading the code, while studying this function
to understand #4301 (for 5b5261401.)  Kind of hard to see how this
way of writing it could ever have seemed like the right thing to do;
it's been this way essentially forever, since f1b10f42a in 2016.

Sadly we can't use the sender's actual, permanent, numeric user ID
here, because outbox messages (values of type Outbox) don't have a
sender_id.  We'll fix that next.  For now, using the email suffices
to fix this bug, and leaves this code no worse than a lot of the
rest of our code.

Fixes: #4303