screen types [nfc]: Rewrite `navigation` type for screens on AppNavigator.
This is NFC because there are no changes to our effective type
annotations -- we just restructure our code to be more like the new
way that's documented [1] for React Navigation v5: alongside each
navigator, have a "param list" type (here, `AppNavigatorParamList`)
that maps a route name to that route's params and a generic helper
type (here, `AppNavigationProp`) that consults it, then have each
screen use that helper to type its `navigation` prop.

As noted in a recent commit where we added src/nav/globalTypes, we
might switch to a different structure after the upgrade is complete,
but we're sticking with the docs for now.

[1] https://reactnavigation.org/docs/typescript