ChatScreen: Stop using @react-navigation/compat.
The `withNavigationFocus` HOC was dropped in v5. Also in v5, they
introduced a `useIsFocused` Hook. But, as Greg points out [1],
they've mitigated the disruption of this sudden change by supplying
`withNavigationFocus` in the compat package, which we've been using.

We made the switch to the new API easier, in a recent commit, by
changing `ChatScreen` from a class component to a function component
that uses React Hooks.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/React.20Navigation.20v5.3A.20compat/near/1058117