types: Remove a dubious type.
I'm not sure why 4d3d6ed408 added this. I don't see any promising
results when I Google React Navigation and "isFocussed" -- with the
"s" doubled -- and the libdef has it as just `focused` in its
`NavigationTabScreenOptions` type.

It seems like there is a problem in the libdef: the `props` param
(from which we're removing this annotation) ends up being inferred
as `empty`, which sounds like some kind of contradiction has been
derived. The `NavigationTabScreenOptions` type from the libdef isn't
even used.

Anyway, we're about to upgrade to React Navigation 5, which will
have better types--that contradict this custom type--and I think it
makes sense to remove an unchecked assumption that isn't obviously
helping us.