flipper ios: Take a RN change that depends on Flipper 41; take Flipper 41.
Part of the RN v0.62 -> v0.63 changes to the template app [1],
corresponding to facebook/react-native@29639e7b9, plus necessary
changes to the Podfile.

The interesting change to the Podfile is the upgrade to Flipper
v0.41.0; the s/DoubleConversion/Flipper-DoubleConversion change is
NFC.

Here's why this commit touches our Podfile, while the RN upstream
commit doesn't: the upstream commit came after
facebook/react-native@619d5d60d, in which React Native stopped
pushing lots of Flipper configuration code out to the consumer's
Podfile. So we'll eventually get to stop making changes like these
to the Podfile, but not until we've taken the changes in React
Native, which we'll do by upgrading to v0.63.

facebook/react-native@29639e7b9 also has a chunk of deletions
alongside the Flipper v0.41.0 upgrade. That chunk is a clean
reversion of a chunk of insertions in
facebook/react-native@27ccc6019 -- so we simplify by not taking the
insertions in the first place.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.62.2&to=0.63.4