recipient [nfc]: Start making the display_recipient invariant explicit.
I.e., the invariant that display_recipient means one thing when
message.type is 'stream', and another when 'private'.

With this change we also provide more-descriptive names in many of
these spots, like `streamName` for the name of a stream.

After this change and the series that follows it, we'll (a) never
consume `display_recipient` without making explicit whether we think
we're looking at a PM or a stream message, and (b) have a real type on
it so that if we did ever consume it directly, the type-checker can
remind us to explicitly confront the wrinkly type it has.