recipient [nfc]: Make the display_recipient invariant explicit everywhere.
This follows up on the commit earlier in this series where we added
these helpers `streamNameOfStreamMessage` and `recipientsOfPrivateMessage`
and started using them everywhere that Flow could detect we were
implicitly relying on an assumption about which kind of display_recipient
we had.  In this commit, we cover everywhere else.

Now we never mention this property at all, except in the type itself;
where we construct values with it; and in those two encapsulating
helpers in recipient.js.  Well, and in comments.  :-)