narrow tests: Fix input data in getNarrowFromMessage test cases.
These recipient lists never happen in real life -- display_recipient
on a PM always contains *all* users involved, including self.  As a
result, we weren't testing the PM scenarios that actually happen in
real life.

While here, also exercise the self-1:1 case, as that's a common
source of bugs in this kind of logic.