recipient: Add error logging for malformed-email fudges in "normalize".
These fudges really don't belong here at all.  For now, just start
logging if they happen, so we can try to confirm that they never do.
This function has a lot of call stacks that get their data from a lot
of different sources, so it's complex to try to confirm that by
studying source code.

The test changes suppress the errors from actually printing in the
normal way (to avoid meaningless noise in the test results), and then
also confirm that the errors were indeed emitted.