recipient: Fix normalizeRecipientsAsUserIds to sort properly.
This code was falling into a classic JS pitfall: calling `Array#sort`
on an array of numbers, with the default comparison.  That comparison
is by *string*, so e.g. 22 sorts before 3 instead of after.

Fixing this bug changes the return value of this function and its
relative normalizeRecipientsAsUserIdsSansMe.  The point of these
functions is to use the return values as keys in our data structures,
so it's important that we're consistent about them.  But as it
happens, the only callers are in the reducer and selector for the
typing status in `state.typing`.  They only need to agree with each
other, so making the change in this one place suffices; and that's
one of the `discardKeys` we don't persist between runs of the app,
so there's no need for a migration.