narrow [nfc]: Add constructor taking Flow-verified proper filtered PM data.
The type on this function's argument, an "opaque type alias",
guarantees that the data comes from one of our helper functions in
`recipient.js`.  See jsdoc within.

This is one of a series of commits which will ultimately let us ensure
that all our PM narrows are constructed with data that (a) observes
the right convention for whether to include the self user, (b) is
sorted consistently, and (c) contains user IDs as well as emails, so
that we can subsequently switch to using user IDs in the Narrow values
themselves.  To enlist Flow to help us with (a) and (b), we'll use
these opaque types to ensure that the data always passes through one
of a few helper functions responsible for those conditions.