presence: Use complete users data structure, from getAllUsersByEmail.
This can only make a difference in the behavior if we have presence
data for a user who's either deactivated or a cross-realm bot.  I'm
not sure that can ever happen... but if we do have such data, it seems
clear that we ought to show it.  So eliminate this one of the few
remaining spots where we used the temptingly-named but incomplete
data from the `getUsers*` selectors.