recipient: Almost always sort recipient lists, by user ID.
This is "almost" because we don't do so in the deprecated
`normalizeRecipientsSansMe`, which only takes emails as input.
There's also still one codepath (through CreateGroupScreen) where
we construct a PM narrow without going through the recipient
helpers at all.

In upcoming commits we'll make this rule universal for constructing
PM narrows.