narrow [nfc]: Distinguish ApiNarrow type, with an alias for now.
We've been taking the data structure that's used in a couple of spots
in the Zulip API as our internal representation of a narrow throughout
the app.  It's a pretty bad fit for that purpose, and we're going to
change to a more reasonable one.  Lay the groundwork by marking the
distinction between the handful of places where we actually mean the
API-facing narrow type, and the rest where we want our own internal
narrow type.

For the present this is just a type alias, because lots of places in
our code still intimately depend on the details of this
representation.

In the upcoming series of commits, we'll fix those so that they all
use explicit interfaces provided by the `narrow` module, after which
we can switch to an opaque type alias so that Flow confirms that
nothing outside that module still depends on the specifics of the
type.  That will then set us up for future changes that actually
alter the type.