narrow [nfc]: Explicitly convert to ApiNarrow where appropriate.
This conversion is a no-op.  But we'll shortly be converting `Narrow`
into an opaque type -- which will make it a no-op that only the
`narrow` module can perform.  That means Flow will check for us that
we've explicitly marked all the places where we intend this
conversion to happen; these are those places.