compose [nfc]: Use IDs, not emails, to compute placeholder.
This could conceivably affect behavior if we're showing a 1:1
conversation with a user we lack data for by email, but have it by
user ID, or vice versa.  We'd switch from showing the generic
"Type a message" to a specific "Message Greg Price" or back.

That shouldn't be possible, though, because the UI to show the
conversation in the first place would have already errored out by
that point.