narrow: Stop storing emails!
This completes a major objective of the long string of refactoring
that appeared in the series of PRs #4330, #4332, #4335, #4339, #4342,
then #4346, #4356, #4361, #4364, and most recently #4368.  After this
change, the portion of #4333 that's about PMs, emails, and user IDs --
aka the portion of #3764 that's about narrows -- is complete.

Still open from #4333 is to convert stream and topic narrows from
stream names to stream IDs.  I'm hopeful that that will be simpler:
(a) unlike the situation for PMs, there's just one stream mentioned
at a time, so there's no question of sorting, and (b) there isn't
the "include self or not" complication that's bedeviled much of our
code related to PMs.  In other words, stream and topic narrows
don't suffer from #4035.