api types: Add UserId, an opaque type alias; don't use it yet.
We'll use this in waves over the coming series of commits.

Because it's public that `UserId` is a subtype of `number`, other code
can freely take a `UserId` and use it as if it were simply a `number`.

That means we can do the conversion gradually so long as we start with
places where a user ID is made, and follow the data flow to places
where it's passed to and ultimately consumed: while we're in the
middle of the conversion, the boundary will always look like some code
providing a `UserId` and some other code accepting it as a `number`.