recipient: Use IDs in isSameRecipient; cut out normalizeRecipients!
The hairy logic in normalizeRecipients was one of the last few cases
of the confusion around PM recipients we're seeking to remove for #4035.

It also represents one of the many places we're in the process of
converting from using emails to using numeric user IDs.

The case where this could potentially change behavior is where we
have several messages that are in the same PM conversation, but the
`display_recipient` property on different messages disagrees about
the email of some participant (e.g. because that user changed their
email between when we learned about one message and when we learned
about the other.)  The old code would fail to notice the messages
were in the same conversation, and so when showing them
consecutively in an interleaved narrow (like all-messages or a
search result), we'd show a new recipient bar between them.

With this fix, we correctly identify the messages as belonging to
the same conversation.