PresenceStatusIndicator: Give presence status icons some contrast.
Now, callers can pass `true` for `useOpaqueBackground`, and there
will be a white background covering the space behind each of the
four status icons (and extending just a little bit beyond, to make a
border).

Use the new feature in `UserAvatarWithPresence`, where the status
icon is overlayed in the lower right corner of the avatar image.

Fixes: #4306

[chris: revised the commit message]