isInternalLink [nfc]: Expand out logic super explicitly.
This isn't a particularly reasonable way to write this function.

But it *is* very easy to confirm behaves exactly the same as the
old version, if you consult the definition of `isUrlOnRealm`.

And splitting the different cases out in this super explicit way
will make it easier to simplify out the complicated parts, and
then to fix the broken parts.