isInternalLink: Fix one of the silly shortcuts here, namely split/pop.
The point of this logic seems to have been to strip off the realm URL
as a prefix if it was there, and leave the URL alone if it wasn't.

It didn't actually do that reliably: if for whatever reason there
happened to be a perfectly legitimate appearance of the realm URL
as a substring *after* the start of the given URL, like:

  https://chat.example.org/foo/https://chat.example.org

or (and this one is a 100% realistic example! try it with your
favorite realm!):

  https://web.archive.org/web/*/https://chat.example.org/terms

then this would end up taking only the part after the last occurrence.
There's no way that behavior can lead to anything good.

It's not really any harder to just write this logic correctly.
So do that.