isInternalLink: Replace with simple `new URL` use; add many tests.
This code sure is simpler than before!

This change causes us to accept a variety of input URLs we
previously didn't: some which the URL Standard defines as invalid
(but nevertheless specifies should be successfully parsed), and
others which are valid and simply make use of some of the more
complicated ways that URL strings can vary while parsing to the same
thing.  I think it's basically fine that we didn't previously accept
these, but it's nice that we do now.

We add a long list of new test cases, almost all of which failed
before this change.  These are perhaps overkill for this function...
but they double as a record of the many tricky features of URLs that
I've learned about from studying the spec in the last couple of days.
A lot of these can be reused, with a bit of adaptation, for testing
other URL-processing functions.

Using `new URL` does have a performance cost, as we've recently
learned.  It's fine for doing a handful of times per user interaction,
though -- and in fact all call stacks here lead to `messageLinkPress`,
which calls it a couple of times upon the user trying to follow a link
in a message, so that's OK.  Propagate a warning in jsdoc up the call
stacks, so that we don't accidentally add a call to these in a
different kind of context.