redux: Provide the global state to each sub-reducer.
There are a lot of places in our various reducers where to best do
their job, they really want some data from elsewhere in our app state,
outside the bit that that particular reducer maintains.

For example, several different reducers want to know when handling an
EVENT_NEW_MESSAGE whether the message was sent by the self user,
because they need to treat it differently in that case.  We currently
accommodate this with a hack, where in `eventToAction` where we turn
the server event into a Redux action, we stick in an extra property
`ownUserId` that's selected straight off the current state.

We have similar hacks to provide `state.caughtUp` on the same event
for the sake of some other reducers, and to provide `ownUserId` or
`state.messages` on some other event actions.  These hacks aren't
limited to actions for server events, either: we have `ownUserId` on
`MESSAGE_FETCH_COMPLETE` actions, too.

And then there's more places where we should be using information
from elsewhere in the state but haven't been, and as a result doing
far more computation than should be necessary to perform some tasks
that ought to be very fast.

For example, when some messages get marked as read -- a very common
operation, sitting in the critical path for updating the "N unreads"
banner, which the user is probably noticing right at that moment --
the "unread" reducer has to sweep through its entire data structure
searching for the given message IDs so it can remove them.  If there
are a lot of unreads, this can take hundreds of ms, which is
painfully long.  What we should be doing there is to consult
`state.messages`, see exactly what conversation each message appears
in, and go straight to that part of the unreads data structure and
touch that and nothing else.

Why haven't we been better using the information available to us in
places like that?  One factor is that it's rather a pain to add one
of these hacks.  Another, probably, is that it's visibly a hack --
and perhaps even because we've felt that the Redux Way is to have
each sub-reducer know nothing about the rest of the state and so
there's something somehow wrong with doing so.

Well, that version of the Redux Way might be a good idea in
situations where different sub-reducers are managing areas of data
that are truly independent of each other, and benefit from being
isolated.  But as these examples show, many or most of our state
subtrees are in reality closely interconnected.  In particular
they're full of IDs that refer to things in other subtrees --
messages, users, streams -- and they're naturally structured by
facts about those things, facts which are maintained in those other
subtrees.  So let's put that idea to rest:

  Our reducers often *should* use data from elsewhere in the state.

  If you think one should, it probably should.

And, with this commit, let's start providing a non-hacky way for
them to do that.  (Or at least a less-hacky way.)

Each sub-reducer is now simply passed the global state as an
argument.  For the moment, they all just ignore it.

Converting any given sub-reducer to actually use that argument will
be a bit of work, just because its tests will all need to be updated
to pass it.

But then once that's done for any given sub-reducer, it can go on to
inspect the state however it pleases, and use any part of the state
that it needs to.