unread [nfc]: Use global state in unreadStreamsReducer, in a small way.
This is one of the places where we've already been using information
from elsewhere in the state, and have done so via a hack where that
information gets stuck into the action as an extra property.

Instead, use our new more-direct mechanism for getting at such data
in a reducer, where it's simply passed down as an argument.

This is a step toward cleaning up that `ownUserId` hack.  It also
serves as a demonstration of the new mechanism, by being the first
place we substantively use it.