types: Use React$ElementConfig rather than $PropertyType<..., 'props'>.
One instance of this, the one in UnreadCards, appeared in a branch
of Ray's.  It's a handy Flow feature that I'm pretty sure I didn't
know about when I first introduced this `$PropertyType<..., 'props'>`
pattern a couple of years ago in 706929bd9.  Let's use it
everywhere it applies.

This also allows it to work properly with the component type
TextInput, so that we get to delete a fixme.  (Sadly the issue
described in the comment by the fixme still exists, as seen at
similar fixmes elsewhere in the codebase; this was actually a
separate problem which just arose at the same time.)  I think the
issue was that in general the old pattern worked only on class
components, not on React component types defined in other ways,
and TextInput is one of the latter.