typing status [nfc]: Factor out pmTypingKeyFromRecipients.
So that the reducer and selector are computing these keys in a way
that clearly matches up.

Also add an implementation comment that explains why they actually
*do* match up.  This version is very straightforwardly doing the
same thing as the old code... but it's not so immediately obvious
that it does the same as `pmTypingKeyFromPmKeyIds`, or that the old
code was doing the same thing in the reducer vs. in the selector.