typing status [nfc]: Mark some parameter arrays as read-only.
This makes these functions a more specific type -- it adds an
additional promise they make to their callers -- so it's compatible
with the existing types of the typing_status library.

The improved types we'll take for that library shortly will require
this promise, so that they can in turn make the same promise to the
caller of `typing_status.update`, a few lines below in this file.