webview: Delay setting scroll position only on iOS.
On iOS/Safari, this logic is essential to avoid landing in a
very wrong scroll position and staying there, #4357.  But on
Android/Chrome, #4357 doesn't happen in the first place; and
introducing this logic causes jank, where we appear to show
messages from the wrong scroll position for one frame.

So, apply this logic only on iOS.