recent-pms: Add a model file, including state type and reducer.
This takes the data from `recent_private_conversations` (which we
just started requesting in the previous commit) in the initial fetch,
and then maintains it as new messages arrive.

We make good use of our new support for Immutable.js data structures.
Instead of maintaining the data in the exact same format as it came
over the wire as JSON, we arrange it on receipt into a data structure
that's optimized for efficiently maintaining and using.

The organization of the code is slightly different from our usual
existing practice:

 * The state type is defined right in the same file as the reducer
   which updates it.

 * Also in that file are some related stateless helper functions:
   in particular `usersOfKey`, which code elsewhere in the app can
   use to help interpret the data provided by this model.

 * The file is named in a general way like "thingModel", rather than
   our usual "thingReducer", and the reducer is a named rather than
   default export.

I think this is helpful here for keeping together code that
logically goes together in that it has shared internal knowledge
and invariants, like the format of PmConversationKey strings.

I think something similar would be helpful for a lot of our model
code (including most of what's in reducers and selectors), but I'm
in no rush to do a sweeping refactor; we'll want to experiment first
and find the arrangement that works best.

This version doesn't put any selector code in this file.  The
existing selector in the neighboring `pmConversationsSelectors.js`
doesn't really belong here, because it mixes together a number of
different areas of the app model, notably the unread counts.  After
some future refactoring to separate that out, there may be a more
focused selector that naturally lives here... or it may be that no
selector is needed, because it'd just be accessing a property on the
state type, and we can simplify by just saying that directly.