recipient: Well-type tests; fix a test; note a bug the test finds!
If we'd had type-checking here in the first place, this test would
have had user IDs in it, and wouldn't have been defeated when we
stopped paying attention to the emails.  Then when we stopped sorting
here, we'd have noticed we were doing so and that we were introducing
this bug.

For the reason explained in the comment, I'd rather not just
reintroduce sorting to fix this.  The bug is low-severity and appears
only in uncommon usage, so for now we just tolerate it.