recipient: Fix pmKeyRecipientsFromIds on empty input; add some tests.
This has been a latent bug: the function wasn't living up to its
documented spec.  I believe it isn't a live one, in that we never
actually pass it an input that triggers this -- all the
representations for PM conversations that we deal with so far
represent the self-1:1 thread with a singleton list, not empty.

We're about to start handling the `recent_private_conversations` data
structure, though, which does use an empty list of users for the
self-1:1 thread.  That turned up this bug; fix it.

Also add some tests for this function, since it evidently needed them.