UserAvatar: Fail gently on bug where auth info is missing.
This fixes #4388 with a quick workaround.  The underlying issue is
still present -- we shouldn't have this component mounted at all in
this situation.  But fixing that may involve some tricky changes to
our navigation, and this is a high-severity issue -- a crash in an
essential flow for using the app for many people -- so we want to fix
it promptly even if the underlying fix takes some time.

Fixes: #4388