recent-pms: Optimize the case of a new message in the latest thread.
This seems likely to be a common case -- probably the majority of
the EVENT_NEW_MESSAGE events.  We already had it taking constant
time (aka O(1) time) thanks to the `i === 0` optimization in
`insertSorted`... but we can do better by not mutating `sorted`
at all.

Also add a test that this optimization is present.  In general we
can't easily test the performance, or the asymptotic complexity, of
our data structures and algorithms... but when the intended behavior
is "the new thing is identical to the old thing", that we can.