a11y: Add "Navigate up" label on up/back button.
Through tests on system settings app and system messages app on
Android, a user hears Navigate up when they click on back button.
So I used this as a label on back button.

For iOS the label should probably be something else, but this is
better than not having it.